Added support for searching and playing youtube playlists
It is now possible to search and queue youtube playlists
without have to provide the playlist url/identifier.

By adding the '-p' flag when invoking the play command,
a youtube playlist search will be performed and the first
match will be queued. If no such match was found, it will
do a basic youtube search command and queue the songs found.

The '-p' command can also be provided in the search command.

Also lowered the output volume to 0.5